pyproject.toml
[tool.poetry]
name = "nonebot-plugin-mahjong-utils"
version = "0.2.7"
description = ""
authors = ["ssttkkl <huang.wen.long@hotmail.com>"]
readme = "README.MD"
license = "MIT"
repository = "https://github.com/ssttkkl/nonebot-plugin-mahjong-utils"
packages = [
    { include = "nonebot_plugin_mahjong_utils", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = "^2.0.0rc1"
mahjong-utils = "^0.4.0"
nonebot-plugin-htmlrender = { version = "^0.2.0.3", optional = true }
nonebot-plugin-send-anything-anywhere = { version = "^0.2.0", optional = true }

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
nb-cli = "^0.6.7"
nonebot-adapter-console = "^0.2.1"
nonebot-adapter-onebot = "^2.2.1"

[tool.poetry.extras]
htmlrender = ["nonebot-plugin-htmlrender", "nonebot-plugin-send-anything-anywhere"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
